Extracted common methods in displaycalendars to CalendarFrame Renamed CalendarFrame to DisplayCalendarHelper
