+++ src/main/java/helpers/DisplayCalendarHelper.java
@@ -3,11 +3,14 @@ package helpers;
 import entities.Event;
 import usecases.EventManager;
 
+import java.time.DayOfWeek;
+import java.time.LocalDate;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 
-public class CalendarFrame {
+public class DisplayCalendarHelper {
 
     private final ArrayList<String> DATES = new ArrayList<>() {
         {
@@ -24,7 +27,7 @@ public class CalendarFrame {
     private final int year;
     private final int month;
 
-    public CalendarFrame(int year, int month){
+    public DisplayCalendarHelper(int year, int month){
         this.year = year;
         this.month = month;
     }
@@ -84,7 +87,71 @@ public class CalendarFrame {
         List<Integer> keyList = new ArrayList<>(calendarMap.keySet());
         for (Integer key:keyList){
             List<Event> eventList = em.timeOrder(calendarMap.get(key));
-            calendarMap.get(key).addAll(eventList);
+            calendarMap.put(key, eventList);
         }
+    }
+
+    public List<String> updateTimeList(List<String> toBeUpdated, List<String> addLst){
+        List<String> temp = new ArrayList<>(toBeUpdated);
+        for (String time : addLst){
+            if (!temp.contains(time)){
+                temp.add(time);
+            }
+        }
+        List<Integer> container = new ArrayList<>();
+        List<String> sortedList = new ArrayList<>();
+        for (String item : temp){
+            container.add(convertTimeToInt(item));
+        }
+        Collections.sort(container);
+        for (Integer number : container) {
+            String convertedTime = String.valueOf(number);
+            if (convertedTime.length() == 1) {
+                sortedList.add("00:0" + convertedTime);
+            } else if (convertedTime.length() == 2) {
+                sortedList.add("00:" + convertedTime);
+            } else if (convertedTime.length() == 3) {
+                sortedList.add("0" + convertedTime.charAt(0) + ":" + convertedTime.substring(1, 3));
+            } else if (convertedTime.length() == 4) {
+                sortedList.add(convertedTime.substring(0, 2) + ":" + convertedTime.substring(2, 4));
+            }
+        }
+        return sortedList;
+    }
+
+    public Integer convertTimeToInt(String time){
+        String temp = time.substring(0, 2) + time.substring(3, 5);
+        return Integer.parseInt(temp);
+    }
+
+    public String findStartDayOfWeek(int year, int month, int date){
+        LocalDate localDate = LocalDate.of(year, month, date);
+        DayOfWeek dayOfWeek = DayOfWeek.from(localDate);
+        int startingDayOfWeek = dayOfWeek.getValue();
+        String chosenDayOfWeek = "Monday";
+        switch (startingDayOfWeek) {
+            case 1:
+                chosenDayOfWeek = "MONDAY";
+                break;
+            case 2:
+                chosenDayOfWeek = "TUESDAY";
+                break;
+            case 3:
+                chosenDayOfWeek = "WEDNESDAY";
+                break;
+            case 4:
+                chosenDayOfWeek = "THURSDAY";
+                break;
+            case 5:
+                chosenDayOfWeek = "FRIDAY";
+                break;
+            case 6:
+                chosenDayOfWeek = "SATURDAY";
+                break;
+            case 7:
+                chosenDayOfWeek = "SUNDAY";
+                break;
+        }
+        return chosenDayOfWeek;
     }
 }
